Support the membase backoff response value.
When membase is under memory pressure and has not yet migrated
enough things to disk to accept new incoming ones, it will send
either a binary or an ascii response telling the client that
there is a temporary failure to accept the store.  At this time
it's appropriate to backoff for a period of time.

This implements a very simple backoff.  A better one would have
some randomness for the interval.  This gets the job done.